+++ include/seqan3/search/fm_index/bi_fm_index_cursor.hpp
@@ -34,6 +34,7 @@ namespace seqan3
 /*!\brief The SeqAn Bidirectional FM Index Cursor.
  * \implements seqan3::bi_fm_index_cursor_specialisation
  * \tparam index_t The type of the underlying index; must model seqan3::bi_fm_index_specialisation.
+ * \implements seqan3::cerealisable
  * \details
  *
  * The cursor's interface provides searching a string both from left to right as well as from right to left in the
@@ -1022,6 +1023,28 @@ public:
                    return locate_result_value_type{sequence_rank-1, sequence_position};
                });
     }
+
+    /*!\cond DEV
+     * \brief Serialisation support function.
+     * \tparam archive_t Type of `archive`; must satisfy seqan3::cereal_archive.
+     * \param archive The archive being serialised from/to.
+     *
+     * \attention These functions are never called directly, see \ref serialisation for more details.
+     */
+    template <cereal_archive archive_t>
+    void CEREAL_SERIALIZE_FUNCTION_NAME(archive_t & archive)
+    {
+        archive(fwd_lb);
+        archive(fwd_rb);
+        archive(rev_lb);
+        archive(rev_rb);
+        archive(sigma);
+        archive(parent_lb);
+        archive(parent_rb);
+        archive(_last_char);
+        archive(depth);
+    }
+    //!\endcond
 };
 
 //!\}